refactor: hooks, lib 폴더 utils 하부로 이동
